The_Django_Website/RedditFlair/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
joblib==0.14.1
nltk==3.5
numpy==1.18.3
pandas==1.0.3
praw==6.5.1
prawcore==1.0.1
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tqdm==4.45.0
update-checker==0.16
urllib3==1.25.9
websocket-client==0.57.0
whitenoise==5.0.1
